+++ src/ralph_assets/forms.py
@@ -534,10 +534,15 @@ class DeviceForm(ModelForm):
     class Meta:
         model = DeviceInfo
         fields = (
-            'u_level',
-            'u_height',
+            'data_center',
+            'server_room',
+            'rack',
+            'orientation',
+            'position',
+            'slot_no',
             'ralph_device_id',
         )
+
     force_unlink = BooleanField(required=False, label=_('Force unlink'))
     create_stock = BooleanField(
         required=False,
@@ -1348,10 +1353,39 @@ class EditDeviceForm(BaseEditAssetForm):
             )
         return cleaned_data
 
+    def clean_hostname(self):
+        # make field readonly
+        return self.instance.hostname or None
+
 
 class BackOfficeEditDeviceForm(ReadOnlyFieldsMixin, EditDeviceForm):
 
-    readonly_fields = ('created',)
+    fieldsets = OrderedDict([
+        ('Basic Info', [
+            'type', 'category', 'model', 'niw', 'barcode', 'sn', 'imei',
+            'warehouse', 'location', 'status', 'task_url', 'loan_end_date',
+            'purpose', 'remarks', 'service_name', 'property_of', 'hostname',
+            'created', 'service', 'device_environment', 'region',
+        ]),
+        ('Financial Info', [
+            'order_no', 'invoice_date', 'invoice_no', 'price', 'provider',
+            'deprecation_rate', 'source', 'request_date',
+            'provider_order_date', 'delivery_date',
+            'deprecation_end_date', 'budget_info', 'force_deprecation',
+        ]),
+        ('User Info', [
+            'user', 'owner', 'employee_id', 'company', 'department', 'manager',
+            'profit_center', 'cost_center', 'segment',
+        ]),
+        ('Assigned licenses info', ['licences']),
+        ('Assigned supports info', [
+            'user', 'owner', 'employee_id', 'company', 'department', 'manager',
+            'profit_center', 'cost_center',
+        ]),
+        ('Assigned supports info', ['required_support', 'supports']),
+    ])
+
+    readonly_fields = ('created', 'hostname')
 
     class Meta(BaseEditAssetForm.Meta):
         fields = BaseEditAssetForm.Meta.fields + (
@@ -1363,9 +1397,6 @@ class BackOfficeEditDeviceForm(ReadOnlyFieldsMixin, EditDeviceForm):
         required=False,
         queryset=models_device.DeviceEnvironment.objects.all(),
         label=_('Environment'),
-    )
-    hostname = CharField(
-        required=False, widget=SimpleReadOnlyWidget(),
     )
     purpose = ChoiceField(
         choices=[('', '----')] + models_assets.AssetPurpose(),
@@ -1385,31 +1416,45 @@ class BackOfficeEditDeviceForm(ReadOnlyFieldsMixin, EditDeviceForm):
 
     def __init__(self, *args, **kwargs):
         super(BackOfficeEditDeviceForm, self).__init__(*args, **kwargs)
-        for after, field in (
-            ('sn', 'imei'),
-            ('loan_end_date', 'purpose'),
-            ('property_of', 'hostname'),
-            ('hostname', 'service'),
-            ('service', 'device_environment'),
-            ('hostname', 'created'),
-        ):
-            self.fieldsets['Basic Info'].append(field)
-            move_after(self.fieldsets['Basic Info'], after, field)
-        self.fieldsets['User Info'].append('segment')
-
-    def clean_hostname(self):
-        # make field readonly
-        return self.instance.hostname or None
+        self.fieldsets = BackOfficeEditDeviceForm.fieldsets
 
 
-class DataCenterEditDeviceForm(EditDeviceForm):
+class DataCenterEditDeviceForm(ReadOnlyFieldsMixin, EditDeviceForm):
+
+    fieldsets = OrderedDict([
+        ('Basic Info', [
+            'type', 'category', 'model', 'niw', 'barcode', 'sn', 'warehouse',
+            'location', 'status', 'task_url', 'loan_end_date', 'remarks',
+            'service_name', 'property_of', 'hostname', 'service',
+            'device_environment', 'region', 'management_ip',
+        ]),
+        ('Financial Info', [
+            'order_no', 'invoice_date', 'invoice_no', 'price', 'provider',
+            'deprecation_rate', 'source', 'request_date',
+            'provider_order_date', 'delivery_date', 'deprecation_end_date',
+            'budget_info', 'force_deprecation',
+        ]),
+        ('User Info', [
+            'user', 'owner', 'employee_id', 'company', 'department', 'manager',
+            'profit_center', 'cost_center',
+        ]),
+        ('Assigned licenses info', ['licences']),
+        ('Assigned supports info', [
+            'user', 'owner', 'employee_id', 'company', 'department', 'manager',
+            'profit_center', 'cost_center',
+        ]),
+        ('Assigned supports info', ['required_support', 'supports']),
+    ])
+
+    readonly_fields = ('hostname',)
 
     edit_management_ip = True
 
     class Meta(BaseEditAssetForm.Meta):
         fields = BaseEditAssetForm.Meta.fields + (
-            'device_environment', 'service',
+            'device_environment', 'service', 'hostname',
         )
+
     device_environment = ModelChoiceField(
         required=True,
         queryset=models_device.DeviceEnvironment.objects.all(),
@@ -1423,12 +1468,12 @@ class DataCenterEditDeviceForm(EditDeviceForm):
 
     def __init__(self, *args, **kwargs):
         super(DataCenterEditDeviceForm, self).__init__(*args, **kwargs)
-        for after, field in (
-            ('property_of', 'service'),
-            ('service', 'device_environment'),
-        ):
-            self.fieldsets['Basic Info'].append(field)
-            move_after(self.fieldsets['Basic Info'], after, field)
+        self.fieldsets = DataCenterEditDeviceForm.fieldsets
+        try:
+            device_hostname = kwargs['instance'].linked_device.name
+        except AttributeError:
+            device_hostname = ''
+        self.initial['hostname'] = device_hostname
 
 
 class SearchAssetForm(Form):